separate sequential interfaces
