+++ include/mockturtle/algorithms/aqfp/aqfp_resynthesis.hpp
@@ -174,30 +174,33 @@ public:
       fanout_resyn_fn( ntk_topo, n, fanouts[n], ntk_dest, node2new[n], 0u, fanout_node_callback, fanout_po_callback ); } );
 
     /* map register outputs */
-    ntk_src.foreach_ro( [&]( auto n )
-                        {
-      auto ro = ntk_dest.create_ro();
-      node2new[n] = ro;
-      level_of_node[ntk_dest.get_node( ro )] = 0u;
+    if constexpr ( has_foreach_ro_v<NtkSrc> && has_create_ro_v<NtkDest> )
+    {
+      ntk_src.foreach_ro( [&]( auto n, auto i )
+                          {
+        auto ro = ntk_dest.create_ro();
+        node2new[n] = ro;
+        level_of_node[ntk_dest.get_node( ro )] = 0u;
 
-      ntk_dest._storage->latch_information[ntk_dest.get_node( node2new[n] )] = ntk_src._storage->latch_information[n];
-      if constexpr ( has_has_name_v<NtkSrc> && has_get_name_v<NtkSrc> && has_set_name_v<NtkDest> )
-      {
-        if ( ntk_src.has_name( ntk_src.make_signal( n ) ) )
-          ntk_dest.set_name( node2new[n], ntk_src.get_name( ntk_src.make_signal( n ) ) );
-      }
+        ntk_dest.set_register( i, ntk_src.register_at( i ) );
+        if constexpr ( has_has_name_v<NtkSrc> && has_get_name_v<NtkSrc> && has_set_name_v<NtkDest> )
+        {
+          if ( ntk_src.has_name( ntk_src.make_signal( n ) ) )
+            ntk_dest.set_name( node2new[n], ntk_src.get_name( ntk_src.make_signal( n ) ) );
+        }
 
-      auto fanout_node_callback = [&]( const auto& f, const auto& level ) {
-        level_for_fanout[{ n, f }] = level;
-      };
+        auto fanout_node_callback = [&]( const auto& f, const auto& level ) {
+          level_for_fanout[{ n, f }] = level;
+        };
 
-      auto fanout_po_callback = [&]( const auto& index, const auto& level ) {
-        (void)index;
-        auto node = ntk_dest.get_node(node2new[n]);
-        po_level_of_node[node] = std::max(po_level_of_node[node], level);
-      };
+        auto fanout_po_callback = [&]( const auto& index, const auto& level ) {
+          (void)index;
+          auto node = ntk_dest.get_node(node2new[n]);
+          po_level_of_node[node] = std::max(po_level_of_node[node], level);
+        };
 
-      fanout_resyn_fn( ntk_topo, n, fanouts[n], ntk_dest, node2new[n], 0u, fanout_node_callback, fanout_po_callback ); } );
+        fanout_resyn_fn( ntk_topo, n, fanouts[n], ntk_dest, node2new[n], 0u, fanout_node_callback, fanout_po_callback ); } );
+    }
 
     /* map nodes */
     ntk_topo.foreach_node( [&]( auto n )
@@ -276,20 +279,23 @@ public:
       } } );
 
     /* map register inputs */
-    ntk_src.foreach_ri( [&]( auto const& f, auto index )
-                        {
-      (void)index;
+    if constexpr ( has_foreach_ri_v<NtkSrc> && has_create_ri_v<NtkDest> )
+    {
+      ntk_src.foreach_ri( [&]( auto const& f, auto index )
+                          {
+        (void)index;
 
-      auto const o = ntk_src.is_complemented( f ) ? ntk_dest.create_not( node2new[f] ) : node2new[f];
-      ntk_dest.create_ri( o );
+        auto const o = ntk_src.is_complemented( f ) ? ntk_dest.create_not( node2new[f] ) : node2new[f];
+        ntk_dest.create_ri( o );
 
-      if constexpr ( has_has_output_name_v<NtkSrc> && has_get_output_name_v<NtkSrc> && has_set_output_name_v<NtkDest> )
-      {
-        if ( ntk_src.has_output_name( index ) )
+        if constexpr ( has_has_output_name_v<NtkSrc> && has_get_output_name_v<NtkSrc> && has_set_output_name_v<NtkDest> )
         {
-          ntk_dest.set_output_name( index + ntk_src.num_pos(), ntk_src.get_output_name( index + ntk_src.num_pos() ) );
-        }
-      } } );
+          if ( ntk_src.has_output_name( index ) )
+          {
+            ntk_dest.set_output_name( index + ntk_src.num_pos(), ntk_src.get_output_name( index + ntk_src.num_pos() ) );
+          }
+        } } );
+    }
 
     return { level_of_node, po_level_of_node };
   }
