+++ include/mockturtle/algorithms/node_resynthesis.hpp
@@ -113,15 +113,18 @@ public:
       }
       } );
 
-    ntk.foreach_ro( [&]( auto n ) {
-      node2new[n] = ntk_dest.create_ro();
-      ntk_dest._storage->latch_information[ntk_dest.get_node(node2new[n])] = ntk._storage->latch_information[n];
-      if constexpr ( has_has_name_v<NtkSource> && has_get_name_v<NtkSource> && has_set_name_v<NtkDest> )
-      {
-        if ( ntk.has_name( ntk.make_signal( n ) ) )
-          ntk_dest.set_name( node2new[n], ntk.get_name( ntk.make_signal( n ) ) );
-      }
-      } );
+    if constexpr ( has_foreach_ro_v<NtkSource> && has_create_ro_v<NtkDest> )
+    {
+      ntk.foreach_ro( [&]( auto n, auto i ) {
+        node2new[n] = ntk_dest.create_ro();
+        ntk_dest.set_register( i, ntk.register_at( i ) );
+        if constexpr ( has_has_name_v<NtkSource> && has_get_name_v<NtkSource> && has_set_name_v<NtkDest> )
+        {
+          if ( ntk.has_name( ntk.make_signal( n ) ) )
+            ntk_dest.set_name( node2new[n], ntk.get_name( ntk.make_signal( n ) ) );
+        }
+        } );
+    }
 
     /* map nodes */
     topo_view ntk_topo{ntk};
@@ -171,20 +174,23 @@ public:
         }
       } );
 
-    ntk.foreach_ri( [&]( auto const& f, auto index ) {
-        (void)index;
+    if constexpr ( has_foreach_ri_v<NtkSource> && has_create_ri_v<NtkDest> )
+    {
+      ntk.foreach_ri( [&]( auto const& f, auto index ) {
+          (void)index;
 
-        auto const o = ntk.is_complemented( f ) ? ntk_dest.create_not( node2new[f] ) : node2new[f];
-        ntk_dest.create_ri( o );
+          auto const o = ntk.is_complemented( f ) ? ntk_dest.create_not( node2new[f] ) : node2new[f];
+          ntk_dest.create_ri( o );
 
-        if constexpr ( has_has_output_name_v<NtkSource> && has_get_output_name_v<NtkSource> && has_set_output_name_v<NtkDest> )
-        {
-          if ( ntk.has_output_name( index ) )
+          if constexpr ( has_has_output_name_v<NtkSource> && has_get_output_name_v<NtkSource> && has_set_output_name_v<NtkDest> )
           {
-            ntk_dest.set_output_name( index + ntk.num_pos(), ntk.get_output_name( index + ntk.num_pos() ) );
+            if ( ntk.has_output_name( index ) )
+            {
+              ntk_dest.set_output_name( index + ntk.num_pos(), ntk.get_output_name( index + ntk.num_pos() ) );
+            }
           }
-        }
-      } );
+        } );
+    }
 
     return ntk_dest;
   }
