+++ include/mockturtle/io/blif_reader.hpp
@@ -35,16 +35,17 @@
 
 #pragma once
 
+#include "../networks/aig.hpp"
+#include "../networks/sequential.hpp"
+#include "../networks/cover.hpp"
+#include "../traits.hpp"
+
 #include <kitty/kitty.hpp>
 #include <lorina/blif.hpp>
 
 #include <map>
 #include <string>
 #include <vector>
-
-#include "../networks/aig.hpp"
-#include "../networks/cover.hpp"
-#include "../traits.hpp"
 
 namespace mockturtle
 {
@@ -88,11 +89,13 @@ public:
       ntk_.create_po( signals[o] );
     }
 
-    for ( auto const& latch : latches )
+    if constexpr ( has_create_ri_v<Ntk> )
     {
-      auto const lit = std::get<0>( latch );
-      auto signal = signals[lit];
-      ntk_.create_ri( signal );
+      for ( auto const& latch : latches )
+      {
+        auto signal = signals[latch];
+        ntk_.create_ri( signal );
+      }
     }
   }
 
@@ -126,87 +129,87 @@ public:
 
   virtual void on_latch( const std::string& input, const std::string& output, const std::optional<latch_type>& l_type, const std::optional<std::string>& control, const std::optional<latch_init_value>& reset ) const override
   {
-    signals[output] = ntk_.create_ro();
-    if constexpr ( has_set_name_v<Ntk> && has_set_output_name_v<Ntk> )
+    if constexpr ( has_create_ro_v<Ntk> )
     {
-      ntk_.set_name( signals[output], output );
-      ntk_.set_output_name( outputs.size() + latches.size(), input );
-    }
-
-    std::string type = "re";
-    if ( l_type )
-    {
-      switch ( *l_type )
+      std::string type = "re";
+      if ( l_type )
       {
-      case latch_type::FALLING:
-      {
-        type = "fe";
-      }
-      break;
-      case latch_type::RISING:
-      {
-        type = "re";
-      }
-      break;
-      case latch_type::ACTIVE_HIGH:
-      {
-        type = "ah";
-      }
-      break;
-      case latch_type::ACTIVE_LOW:
-      {
-        type = "al";
-      }
-      break;
-      case latch_type::ASYNC:
-      {
-        type = "as";
-      }
-      break;
-      default:
-      {
-        type = "";
-      }
-      break;
-      }
-    }
-
-    uint32_t r = 3;
-    if ( reset )
-    {
-      switch ( *reset )
-      {
-      case latch_init_value::NONDETERMINISTIC:
-      {
-        r = 2;
-      }
-      break;
-      case latch_init_value::ONE:
-      {
-        r = 1;
-      }
-      break;
-      case latch_init_value::ZERO:
-      {
-        r = 0;
-      }
-      break;
-      default:
+        switch ( *l_type )
+        {
+        case latch_type::FALLING:
+        {
+          type = "fe";
+        }
         break;
+        case latch_type::RISING:
+        {
+          type = "re";
+        }
+        break;
+        case latch_type::ACTIVE_HIGH:
+        {
+          type = "ah";
+        }
+        break;
+        case latch_type::ACTIVE_LOW:
+        {
+          type = "al";
+        }
+        break;
+        case latch_type::ASYNC:
+        {
+          type = "as";
+        }
+        break;
+        default:
+        {
+          type = "";
+        }
+        break;
+        }
       }
+
+      uint32_t r = 3;
+      if ( reset )
+      {
+        switch ( *reset )
+        {
+        case latch_init_value::NONDETERMINISTIC:
+        {
+          r = 2;
+        }
+        break;
+        case latch_init_value::ONE:
+        {
+          r = 1;
+        }
+        break;
+        case latch_init_value::ZERO:
+        {
+          r = 0;
+        }
+        break;
+        default:
+          break;
+        }
+      }
+
+      std::string ctrl = control.has_value() ? control.value() : "clock";
+
+      register_t reg;
+      reg.control = ctrl;
+      reg.init = r;
+      reg.type = type;
+
+      signals[output] = ntk_.create_ro( reg );
+      if constexpr ( has_set_name_v<Ntk> && has_set_output_name_v<Ntk> )
+      {
+        ntk_.set_name( signals[output], output );
+        ntk_.set_output_name( outputs.size() + latches.size(), input );
+      }
+      
+      latches.emplace_back( input );
     }
-
-    latch_info l_info;
-    l_info.init = r;
-    l_info.type = type;
-
-    if ( control.has_value() )
-      l_info.control = control.value();
-    else
-      l_info.control = "clock";
-
-    ntk_._storage->latch_information[ntk_.get_node( signals[output] )] = l_info;
-    latches.emplace_back( std::make_tuple( input, r, type, l_info.control, "" ) );
   }
 
   virtual void on_gate( const std::vector<std::string>& inputs, const std::string& output, const output_cover_t& cover ) const override
@@ -319,7 +322,7 @@ private:
 
   mutable std::map<std::string, signal<Ntk>> signals;
   mutable std::vector<std::string> outputs;
-  mutable std::vector<std::tuple<std::string, int8_t, std::string, std::string, std::string>> latches;
+  mutable std::vector<std::string> latches;
 }; /* blif_reader */
 
 } /* namespace mockturtle */
