+++ include/mockturtle/networks/aig.hpp
@@ -202,42 +202,21 @@ public:
     auto& node = _storage->nodes.emplace_back();
     node.children[0].data = node.children[1].data = _storage->inputs.size();
     _storage->inputs.emplace_back( index );
-    ++_storage->num_pis;
     return {index, 0};
   }
 
-  uint32_t create_po( signal const& f)
+  uint32_t create_po( signal const& f )
   {
     /* increase ref-count to children */
     _storage->nodes[f.index].data[0].h1++;
     auto const po_index = _storage->outputs.size();
     _storage->outputs.emplace_back( f.index, f.complement );
-    ++_storage->num_pos;
     return static_cast<uint32_t>( po_index );
-  }
-
-  signal create_ro()
-  {
-    auto const index = _storage->nodes.size();
-    auto& node = _storage->nodes.emplace_back();
-    node.children[0].data = node.children[1].data = _storage->inputs.size();
-    _storage->inputs.emplace_back( index );
-    return {index, 0};
-  }
-
-  uint32_t create_ri( signal const& f )
-  {
-    /* increase ref-count to children */
-    _storage->nodes[f.index].data[0].h1++;
-    auto const ri_index = _storage->outputs.size();
-    _storage->outputs.emplace_back( f.index, f.complement );
-    return static_cast<uint32_t>( ri_index );
   }
 
   bool is_combinational() const
   {
-    return ( static_cast<uint32_t>( _storage->inputs.size() ) == _storage->num_pis &&
-             static_cast<uint32_t>( _storage->outputs.size() ) == _storage->num_pos );
+    return true;
   }
 
   bool is_constant( node const& n ) const
@@ -252,12 +231,7 @@ public:
 
   bool is_pi( node const& n ) const
   {
-    return _storage->nodes[n].children[0].data == _storage->nodes[n].children[1].data && _storage->nodes[n].children[0].data < static_cast<uint64_t>(_storage->num_pis);
-  }
-
-  bool is_ro( node const& n ) const
-  {
-    return _storage->nodes[n].children[0].data == _storage->nodes[n].children[1].data && _storage->nodes[n].children[0].data >= static_cast<uint64_t>(_storage->num_pis);
+    return _storage->nodes[n].children[0].data == _storage->nodes[n].children[1].data && !is_constant( n );
   }
 
   bool constant_value( node const& n ) const
@@ -703,18 +677,12 @@ public:
 
   auto num_pis() const
   {
-    return _storage->num_pis;
+    return static_cast<uint32_t>( _storage->inputs.size() );
   }
 
   auto num_pos() const
   {
-    return _storage->num_pos;
-  }
-
-  auto num_registers() const
-  {
-    assert( static_cast<uint32_t>( _storage->inputs.size() - _storage->num_pis ) == static_cast<uint32_t>( _storage->outputs.size() - _storage->num_pos ) );
-    return static_cast<uint32_t>( _storage->inputs.size() - _storage->num_pis );
+    return static_cast<uint32_t>( _storage->outputs.size() );
   }
 
   auto num_gates() const
@@ -848,26 +816,14 @@ public:
 
   node pi_at( uint32_t index ) const
   {
-    assert( index < _storage->num_pis );
+    assert( index < _storage->inputs.size() );
     return *(_storage->inputs.begin() + index);
   }
 
   signal po_at( uint32_t index ) const
   {
-    assert( index < _storage->num_pos );
+    assert( index < _storage->outputs.size() );
     return *(_storage->outputs.begin() + index);
-  }
-
-  node ro_at( uint32_t index ) const
-  {
-    assert( index < _storage->inputs.size() - _storage->num_pis );
-    return *(_storage->inputs.begin() + _storage->num_pis + index);
-  }
-
-  signal ri_at( uint32_t index ) const
-  {
-    assert( index < _storage->outputs.size() - _storage->num_pos );
-    return *(_storage->outputs.begin() + _storage->num_pos + index);
   }
 
   uint32_t ci_index( node const& n ) const
@@ -893,8 +849,6 @@ public:
   uint32_t pi_index( node const& n ) const
   {
     assert( _storage->nodes[n].children[0].data == _storage->nodes[n].children[1].data );
-    assert( _storage->nodes[n].children[0].data < _storage->num_pis );
-
     return static_cast<uint32_t>( _storage->nodes[n].children[0].data );
   }
 
@@ -910,38 +864,6 @@ public:
         return true;
       });
     return i;
-  }
-
-  uint32_t ro_index( node const& n ) const
-  {
-    assert( _storage->nodes[n].children[0].data == _storage->nodes[n].children[1].data );
-    assert( _storage->nodes[n].children[0].data >= _storage->num_pis );
-
-    return static_cast<uint32_t>( _storage->nodes[n].children[0].data - _storage->num_pis );
-  }
-
-  uint32_t ri_index( signal const& s ) const
-  {
-    uint32_t i = -1;
-    foreach_ri( [&]( const auto& x, auto index ){
-        if ( x == s )
-        {
-          i = index;
-          return false;
-        }
-        return true;
-      });
-    return i;
-  }
-
-  signal ro_to_ri( signal const& s ) const
-  {
-    return *( _storage->outputs.begin() + _storage->num_pos + _storage->nodes[s.index].children[0].data - _storage->num_pis );
-  }
-
-  node ri_to_ro( signal const& s ) const
-  {
-    return *( _storage->inputs.begin() + _storage->num_pis + ri_index( s ) );
   }
 #pragma endregion
 
@@ -970,70 +892,13 @@ public:
   template<typename Fn>
   void foreach_pi( Fn&& fn ) const
   {
-    detail::foreach_element( _storage->inputs.begin(), _storage->inputs.begin() + _storage->num_pis, fn );
+    detail::foreach_element( _storage->inputs.begin(), _storage->inputs.end(), fn );
   }
 
   template<typename Fn>
   void foreach_po( Fn&& fn ) const
   {
-    detail::foreach_element( _storage->outputs.begin(), _storage->outputs.begin() + _storage->num_pos, fn );
-  }
-
-  template<typename Fn>
-  void foreach_ro( Fn&& fn ) const
-  {
-    detail::foreach_element( _storage->inputs.begin() + _storage->num_pis, _storage->inputs.end(), fn );
-  }
-
-  template<typename Fn>
-  void foreach_ri( Fn&& fn ) const
-  {
-    detail::foreach_element( _storage->outputs.begin() + _storage->num_pos, _storage->outputs.end(), fn );
-  }
-
-  template<typename Fn>
-  void foreach_register( Fn&& fn ) const
-  {
-    static_assert( detail::is_callable_with_index_v<Fn, std::pair<signal,node>, void> ||
-                   detail::is_callable_without_index_v<Fn, std::pair<signal,node>, void> ||
-                   detail::is_callable_with_index_v<Fn, std::pair<signal,node>, bool> ||
-                   detail::is_callable_without_index_v<Fn, std::pair<signal,node>, bool> );
-
-    assert( _storage->inputs.size() - _storage->num_pis == _storage->outputs.size() - _storage->num_pos );
-    auto ro = _storage->inputs.begin() + _storage->num_pis;
-    auto ri = _storage->outputs.begin() + _storage->num_pos;
-    if constexpr ( detail::is_callable_without_index_v<Fn, std::pair<signal,node>, bool> )
-    {
-      while ( ro != _storage->inputs.end() && ri != _storage->outputs.end() )
-      {
-        if ( !fn( std::make_pair(ri++, ro++) ) )
-          return;
-      }
-    }
-    else if constexpr ( detail::is_callable_with_index_v<Fn, std::pair<signal,node>, bool> )
-    {
-      uint32_t index{0};
-      while ( ro != _storage->inputs.end() && ri != _storage->outputs.end() )
-      {
-        if ( !fn( std::make_pair(ri++, ro++), index++ ) )
-          return;
-      }
-    }
-    else if constexpr( detail::is_callable_without_index_v<Fn, std::pair<signal,node>, void> )
-    {
-      while ( ro != _storage->inputs.end() && ri != _storage->outputs.end() )
-      {
-        fn( std::make_pair(*ri++, *ro++) );
-      }
-    }
-    else if constexpr ( detail::is_callable_with_index_v<Fn, std::pair<signal,node>, void> )
-    {
-      uint32_t index{0};
-      while ( ro != _storage->inputs.end() && ri != _storage->outputs.end() )
-      {
-        fn( std::make_pair(*ri++, *ro++), index++ );
-      }
-    }
+    detail::foreach_element( _storage->outputs.begin(), _storage->outputs.end(), fn );
   }
 
   template<typename Fn>
