+++ include/mockturtle/networks/klut.hpp
@@ -176,7 +176,6 @@ public:
     _storage->nodes.emplace_back();
     _storage->inputs.emplace_back( index );
     _storage->nodes[index].data[1].h1 = 2;
-    ++_storage->num_pis;
     return index;
   }
 
@@ -184,35 +183,14 @@ public:
   {
     /* increase ref-count to children */
     _storage->nodes[f].data[0].h1++;
-
     auto const po_index = static_cast<uint32_t>( _storage->outputs.size() );
     _storage->outputs.emplace_back( f );
-    ++_storage->num_pos;
     return po_index;
-  }
-
-  signal create_ro()
-  {
-    auto const index = static_cast<uint32_t>( _storage->nodes.size() );
-    _storage->nodes.emplace_back();
-    _storage->inputs.emplace_back( index );
-    _storage->nodes[index].data[1].h1 = 2;
-    return index;
-  }
-
-  uint32_t create_ri( signal const& f )
-  {
-    /* increase ref-count to children */
-    _storage->nodes[f].data[0].h1++;
-    auto const ri_index = static_cast<uint32_t>( _storage->outputs.size() );
-    _storage->outputs.emplace_back( f );
-    return ri_index;
   }
 
   bool is_combinational() const
   {
-    return ( static_cast<uint32_t>( _storage->inputs.size() ) == _storage->num_pis &&
-             static_cast<uint32_t>( _storage->outputs.size() ) == _storage->num_pos );
+    return true;
   }
 
   bool is_constant( node const& n ) const
@@ -227,14 +205,7 @@ public:
 
   bool is_pi( node const& n ) const
   {
-    const auto end = _storage->inputs.begin() + _storage->num_pis;
-
-    return std::find( _storage->inputs.begin(), end, n ) != end;
-  }
-
-  bool is_ro( node const& n ) const
-  {
-    return std::find( _storage->inputs.begin() + _storage->num_pis, _storage->inputs.end(), n ) != _storage->inputs.end();
+    return std::find( _storage->inputs.begin(), _storage->inputs.end(), n ) != _storage->inputs.end();
   }
 
   bool constant_value( node const& n ) const
@@ -433,18 +404,12 @@ signal create_maj( signal a, signal b, signal c )
 
   auto num_pis() const
   {
-    return _storage->num_pis;
+    return static_cast<uint32_t>( _storage->inputs.size() );
   }
 
   auto num_pos() const
   {
-    return _storage->num_pos;
-  }
-
-  auto num_registers() const
-  {
-    assert( static_cast<uint32_t>( _storage->inputs.size() - _storage->num_pis ) == static_cast<uint32_t>( _storage->outputs.size() - _storage->num_pos ) );
-    return static_cast<uint32_t>( _storage->inputs.size() - _storage->num_pis );
+    return static_cast<uint32_t>( _storage->outputs.size() );
   }
 
   auto num_gates() const
@@ -516,96 +481,14 @@ signal create_maj( signal a, signal b, signal c )
 
   node pi_at( uint32_t index ) const
   {
-    assert( index < _storage->num_pis );
+    assert( index < _storage->inputs.size() );
     return *(_storage->inputs.begin() + index);
   }
 
   signal po_at( uint32_t index ) const
   {
-    assert( index < _storage->num_pos );
+    assert( index < _storage->outputs.size() );
     return (_storage->outputs.begin() + index)->index;
-  }
-
-  node ro_at( uint32_t index ) const
-  {
-    assert( index < _storage->inputs.size() - _storage->num_pis );
-    return *(_storage->inputs.begin() + _storage->num_pis + index);
-  }
-
-  signal ri_at( uint32_t index ) const
-  {
-    assert( index < _storage->outputs.size() - _storage->num_pos );
-    return (_storage->outputs.begin() + _storage->num_pos + index)->index;
-  }
-
-  uint32_t ci_index( node const& n ) const
-  {
-    assert( _storage->nodes[n].children[0].data == _storage->nodes[n].children[1].data );
-    return static_cast<uint32_t>( _storage->nodes[n].children[0].data );
-  }
-
-  uint32_t co_index( signal const& s ) const
-  {
-    uint32_t i = -1;
-    foreach_co( [&]( const auto& x, auto index ){
-        if ( x == s )
-        {
-          i = index;
-          return false;
-        }
-        return true;
-      });
-    return i;
-  }
-
-  uint32_t pi_index( node const& n ) const
-  {
-    assert( _storage->nodes[n].children[0].data == _storage->nodes[n].children[1].data );
-    return static_cast<uint32_t>( _storage->nodes[n].children[0].data );
-  }
-
-  uint32_t po_index( signal const& s ) const
-  {
-    uint32_t i = -1;
-    foreach_po( [&]( const auto& x, auto index ){
-        if ( x == s )
-        {
-          i = index;
-          return false;
-        }
-        return true;
-      });
-    return i;
-  }
-
-  uint32_t ro_index( node const& n ) const
-  {
-    assert( _storage->nodes[n].children[0].data == _storage->nodes[n].children[1].data );
-    return static_cast<uint32_t>( _storage->nodes[n].children[0].data - _storage->num_pis );
-  }
-
-  uint32_t ri_index( signal const& s ) const
-  {
-    uint32_t i = -1;
-    foreach_ri( [&]( const auto& x, auto index ){
-        if ( x == s )
-        {
-          i = index;
-          return false;
-        }
-        return true;
-      });
-    return i;
-  }
-
-  signal ro_to_ri( signal const& s ) const
-  {
-    return ( _storage->outputs.begin() + _storage->num_pos + _storage->nodes[s].children[0].data - _storage->num_pis )->index;
-  }
-
-  node ri_to_ro( signal const& s ) const
-  {
-    return *( _storage->inputs.begin() + _storage->num_pis + ri_index( s ) );
   }
 #pragma endregion
 
@@ -633,72 +516,14 @@ signal create_maj( signal a, signal b, signal c )
   template<typename Fn>
   void foreach_pi( Fn&& fn ) const
   {
-    detail::foreach_element( _storage->inputs.begin(), _storage->inputs.begin() + _storage->num_pis, fn );
+    detail::foreach_element( _storage->inputs.begin(), _storage->inputs.end(), fn );
   }
 
   template<typename Fn>
   void foreach_po( Fn&& fn ) const
   {
     using IteratorType = decltype( _storage->outputs.begin() );
-    detail::foreach_element_transform<IteratorType, uint32_t>( _storage->outputs.begin(), _storage->outputs.begin() + _storage->num_pos, []( auto o ) { return o.index; }, fn );
-  }
-
-  template<typename Fn>
-  void foreach_ro( Fn&& fn ) const
-  {
-    detail::foreach_element( _storage->inputs.begin() + _storage->num_pis, _storage->inputs.end(), fn );
-  }
-
-  template<typename Fn>
-  void foreach_ri( Fn&& fn ) const
-  {
-    using IteratorType = decltype( _storage->outputs.begin() );
-    detail::foreach_element_transform<IteratorType, uint32_t>( _storage->outputs.begin() + _storage->num_pos, _storage->outputs.end(), []( auto o ) { return o.index; }, fn );
-  }
-
-  template<typename Fn>
-  void foreach_register( Fn&& fn ) const
-  {
-    static_assert( detail::is_callable_with_index_v<Fn, std::pair<signal,node>, void> ||
-                   detail::is_callable_without_index_v<Fn, std::pair<signal,node>, void> ||
-                   detail::is_callable_with_index_v<Fn, std::pair<signal,node>, bool> ||
-                   detail::is_callable_without_index_v<Fn, std::pair<signal,node>, bool> );
-
-    assert( _storage->inputs.size() - _storage->num_pis == _storage->outputs.size() - _storage->num_pos );
-    auto ro = _storage->inputs.begin() + _storage->num_pis;
-    auto ri = _storage->outputs.begin() + _storage->num_pos;
-    if constexpr ( detail::is_callable_without_index_v<Fn, std::pair<signal,node>, bool> )
-    {
-      while ( ro != _storage->inputs.end() && ri != _storage->outputs.end() )
-      {
-        if ( !fn( std::make_pair( ( ri++ )->index, ro++ ) ) )
-          return;
-      }
-    }
-    else if constexpr ( detail::is_callable_with_index_v<Fn, std::pair<signal,node>, bool> )
-    {
-      uint32_t index{0};
-      while ( ro != _storage->inputs.end() && ri != _storage->outputs.end() )
-      {
-        if ( !fn( std::make_pair( ( ri++ )->index, ro++ ), index++ ) )
-          return;
-      }
-    }
-    else if constexpr( detail::is_callable_without_index_v<Fn, std::pair<signal,node>, void> )
-    {
-      while ( ro != _storage->inputs.end() && ri != _storage->outputs.end() )
-      {
-        fn( std::make_pair( ( ri++ )->index, *ro++ ) );
-      }
-    }
-    else if constexpr ( detail::is_callable_with_index_v<Fn, std::pair<signal,node>, void> )
-    {
-      uint32_t index{0};
-      while ( ro != _storage->inputs.end() && ri != _storage->outputs.end() )
-      {
-        fn( std::make_pair( ( ri++ )->index, *ro++ ), index++ );
-      }
-    }
+    detail::foreach_element_transform<IteratorType, uint32_t>( _storage->outputs.begin(), _storage->outputs.end(), []( auto o ) { return o.index; }, fn );
   }
 
   template<typename Fn>
