+++ test/networks/cover.cpp
@@ -78,122 +78,6 @@ TEST_CASE( "create and use primary outputs in a cover network", "[cover]" )
   CHECK( cover.num_pos() == 3 );
 }
 
-TEST_CASE( "create and use register in a cover network", "[cover]" )
-{
-  cover_network cover;
-
-  CHECK( has_foreach_po_v<cover_network> );
-  CHECK( has_create_po_v<cover_network> );
-  CHECK( has_create_pi_v<cover_network> );
-  CHECK( has_create_ro_v<cover_network> );
-  CHECK( has_create_ri_v<cover_network> );
-  CHECK( has_create_maj_v<cover_network> );
-
-  const auto c0 = cover.get_constant( false );
-  const auto x1 = cover.create_pi();
-  const auto x2 = cover.create_pi();
-  const auto x3 = cover.create_pi();
-  const auto x4 = cover.create_pi();
-
-  CHECK( cover.size() == 6 );
-  CHECK( cover.num_registers() == 0 );
-  CHECK( cover.num_cis() == 4 );
-  CHECK( cover.num_cos() == 0 );
-
-  const auto f1 = cover.create_maj( x1, x2, x3 );
-  cover.create_po( f1 );
-
-  CHECK( cover.num_pos() == 1 );
-
-  const auto s1 = cover.create_ro(); // ntk. input
-  cover.create_po( s1 );             // po
-
-  const auto f2 = cover.create_maj( f1, x4, c0 );
-  cover.create_ri( f2 ); // ntk. output
-
-  CHECK( cover.num_registers() == 1 );
-  CHECK( cover.num_cis() == 4 + 1 );
-  CHECK( cover.num_cos() == 2 + 1 );
-
-  cover.foreach_pi( [&]( auto const& node, auto index ) {
-    CHECK( cover.is_pi( node ) );
-    switch ( index )
-    {
-    case 0:
-      CHECK( node == x1 ); /* first pi */
-      break;
-    case 1:
-      CHECK( node == x2 ); /* second pi */
-      break;
-    case 2:
-      CHECK( node == x3 ); /* third pi */
-      break;
-    case 3:
-      CHECK( node == x4 ); /* fourth pi */
-      break;
-    default:
-      CHECK( false );
-    }
-  } );
-
-  cover.foreach_ci( [&]( auto const& node, auto index ) {
-    CHECK( cover.is_ci( node ) );
-    switch ( index )
-    {
-    case 0:
-      CHECK( node == x1 ); /* first pi */
-      break;
-    case 1:
-      CHECK( node == x2 ); /* second pi */
-      break;
-    case 2:
-      CHECK( node == x3 ); /* third pi */
-      break;
-    case 3:
-      CHECK( node == x4 ); /* fourth pi */
-      break;
-    case 4:
-      CHECK( node == s1 ); /* first state-bit */
-      CHECK( cover.is_ci( node ) );
-      CHECK( !cover.is_pi( node ) );
-      break;
-    default:
-      CHECK( false );
-    }
-  } );
-
-  cover.foreach_po( [&]( auto const& node, auto index ) {
-    switch ( index )
-    {
-    case 0:
-      CHECK( node == f1 ); /* first po */
-      break;
-    case 1:
-      CHECK( node == s1 ); /* second po */
-      break;
-    default:
-      CHECK( false );
-    }
-  } );
-
-  cover.foreach_co( [&]( auto const& node, auto index ) {
-    switch ( index )
-    {
-    case 0:
-      CHECK( node == f1 ); /* first po */
-      break;
-    case 1:
-      CHECK( node == s1 ); /* second po */
-      break;
-    case 2:
-      CHECK( node == f2 ); /* first next-state bit */
-      break;
-    default:
-      CHECK( false );
-    }
-  } );
-}
-
 TEST_CASE( "create unary operations in a cover network", "[cover]" )
 {
   cover_network cover;
